+++ cachalot/api.py
@@ -1,4 +1,5 @@
 from contextlib import contextmanager
+from typing import Any, Optional, Tuple, Union
 
 from django.apps import apps
 from django.conf import settings
@@ -45,7 +46,11 @@ def _get_tables(tables_or_models):
                else table_or_model._meta.db_table)
 
 
-def invalidate(*tables_or_models, **kwargs):
+def invalidate(
+    *tables_or_models: Tuple[Union[str, Any], ...],
+    cache_alias: Optional[str] = None,
+    db_alias: Optional[str] = None,
+) -> None:
     """
     Clears what was cached by django-cachalot implying one or more SQL tables
     or models from ``tables_or_models``.
@@ -62,19 +67,9 @@ def invalidate(*tables_or_models, **kwargs):
                            (or a combination)
     :type tables_or_models: tuple of strings or models
     :arg cache_alias: Alias from the Django ``CACHES`` setting
-    :type cache_alias: string or NoneType
     :arg db_alias: Alias from the Django ``DATABASES`` setting
-    :type db_alias: string or NoneType
     :returns: Nothing
-    :rtype: NoneType
     """
-    # TODO: Replace with positional arguments when we drop Python 2 support.
-    cache_alias = kwargs.pop('cache_alias', None)
-    db_alias = kwargs.pop('db_alias', None)
-    for k in kwargs:
-        raise TypeError(
-            "invalidate() got an unexpected keyword argument '%s'" % k)
-
     send_signal = False
     invalidated = set()
     for cache_alias, db_alias, tables in _cache_db_tables_iterator(
@@ -90,7 +85,11 @@ def invalidate(*tables_or_models, **kwargs):
             post_invalidation.send(table, db_alias=db_alias)
 
 
-def get_last_invalidation(*tables_or_models, **kwargs):
+def get_last_invalidation(
+    *tables_or_models: Tuple[Union[str, Any], ...],
+    cache_alias: Optional[str] = None,
+    db_alias: Optional[str] = None,
+) -> float:
     """
     Returns the timestamp of the most recent invalidation of the given
     ``tables_or_models``.  If ``tables_or_models`` is not specified,
@@ -106,19 +105,9 @@ def get_last_invalidation(*tables_or_models, **kwargs):
                            (or a combination)
     :type tables_or_models: tuple of strings or models
     :arg cache_alias: Alias from the Django ``CACHES`` setting
-    :type cache_alias: string or NoneType
     :arg db_alias: Alias from the Django ``DATABASES`` setting
-    :type db_alias: string or NoneType
     :returns: The timestamp of the most recent invalidation
-    :rtype: float
     """
-    # TODO: Replace with positional arguments when we drop Python 2 support.
-    cache_alias = kwargs.pop('cache_alias', None)
-    db_alias = kwargs.pop('db_alias', None)
-    for k in kwargs:
-        raise TypeError("get_last_invalidation() got an unexpected "
-                        "keyword argument '%s'" % k)
-
     last_invalidation = 0.0
     for cache_alias, db_alias, tables in _cache_db_tables_iterator(
             list(_get_tables(tables_or_models)), cache_alias, db_alias):
@@ -134,7 +123,7 @@ def get_last_invalidation(*tables_or_models, **kwargs):
 
 
 @contextmanager
-def cachalot_disabled(all_queries=False):
+def cachalot_disabled(all_queries: bool = False):
     """
     Context manager for temporarily disabling cachalot.
     If you evaluate the same queryset a second time,
@@ -158,7 +147,6 @@ def cachalot_disabled(all_queries=False):
     the original and duplicate query.
 
     :arg all_queries: Any query, including already evaluated queries, are re-evaluated.
-    :type all_queries: bool
     """
     was_enabled = getattr(LOCAL_STORAGE, "cachalot_enabled", cachalot_settings.CACHALOT_ENABLED)
     LOCAL_STORAGE.cachalot_enabled = False