+++ cachalot/tests/migrations/0001_initial.py
@@ -2,7 +2,7 @@ from django import VERSION as DJANGO_VERSION
 from django.conf import settings
 from django.contrib.postgres.fields import (
     ArrayField, HStoreField, IntegerRangeField,
-    DateRangeField, DateTimeRangeField)
+    DateRangeField, DateTimeRangeField, DecimalRangeField)
 from django.contrib.postgres.operations import (
     HStoreExtension, UnaccentExtension)
 from django.db import models, migrations
@@ -21,19 +21,6 @@ def extra_regular_available_fields():
 
 def extra_postgres_available_fields():
     fields = []
-    try:
-        # TODO Remove when Dj31 support is dropped
-        from django.contrib.postgres.fields import FloatRangeField
-        fields.append(('float_range', FloatRangeField(null=True, blank=True)))
-    except ImportError:
-        pass
-
-    try:
-        # TODO Add to module import when Dj31 is dropped
-        from django.contrib.postgres.fields import DecimalRangeField
-        fields.append(('decimal_range', DecimalRangeField(null=True, blank=True)))
-    except ImportError:
-        pass
 
     # Future proofing with Django 40 deprecation
     if DJANGO_VERSION[0] < 4:
@@ -103,6 +90,7 @@ class Migration(migrations.Migration):
                 ('int_range', IntegerRangeField(null=True, blank=True)),
                 ('date_range', DateRangeField(null=True, blank=True)),
                 ('datetime_range', DateTimeRangeField(null=True, blank=True)),
+                ('decimal_range', DecimalRangeField(null=True, blank=True))
             ] + extra_postgres_available_fields(),
         ),
         migrations.RunSQL('CREATE TABLE cachalot_unmanagedmodel '